add shared libary

+++ learning-engine-service/src/main/java/com/esprit/learning_engine_service/services/servicesImpl/LearningCategoryServiceImpl.java
@@ -4,6 +4,7 @@ import com.esprit.learning_engine_service.entities.LearningCategory;
 import com.esprit.learning_engine_service.repositories.LearningCategoryRepository;
 import com.esprit.learning_engine_service.services.LearningCategoryService;
 import org.springframework.stereotype.Service;
+import tn.esprit.spring.commonlib.exceptions.ResourceNotFoundException;
 
 import java.util.List;
 
@@ -23,7 +24,8 @@ public class LearningCategoryServiceImpl implements LearningCategoryService {
 
     @Override
     public LearningCategory getById(String id) {
-        return repo.findById(id).orElse(null);
+        return repo.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("LearningCategory not found with id: " + id));
     }
 
     @Override
@@ -33,6 +35,9 @@ public class LearningCategoryServiceImpl implements LearningCategoryService {
 
     @Override
     public void delete(String id) {
+        if (!repo.existsById(id)) {
+            throw new ResourceNotFoundException("LearningCategory not found with id: " + id);
+        }
         repo.deleteById(id);
     }
 }
